feat: tidy up apis, add user ownership middleware to decide if the user is allowed for the api request or not

+++ routes/api.php
@@ -37,15 +37,15 @@ Route::get('/clear-cache', function () {
 });
 
 Route::get('/users', [UserController::class, 'getUsers']);
-Route::get('/users/all-users', [UserController::class, 'getAllUsers']);
-Route::get('/users/all-admins', [UserController::class, 'getAllAdmins']);
+Route::get('/users/all-users', [UserController::class, 'getAllUsers'])->middleware(['auth:sanctum', 'room.access']);
+Route::get('/users/all-admins', [UserController::class, 'getAllAdmins'])->middleware(['auth:sanctum', 'admin']);
 
 // Token
 Route::post('/users/login', [UserController::class, 'login']);
 // Route::middleware('auth:sanctum')->group(function () {
 //     Route::get('users/refresh-token', [UserController::class, 'refreshToken']);
 // });
-Route::middleware('auth:sanctum', 'ability:access-api')->group(function () {
+Route::middleware('auth:sanctum')->group(function () {
     Route::get('/user', function (Request $request) {
         return $request->user();
     });
@@ -63,7 +63,7 @@ Route::middleware('auth:sanctum')->group(function () {
 });
 
 // Admin - Deck Routes
-Route::get('/decks/{deck}/origins', [DeckController::class, 'getOrigins']);
+Route::get('/decks/{deck}/origins', [DeckController::class, 'getOrigins'])->middleware(['auth:sanctum', 'admin']);
 
 Route::middleware(['auth:sanctum', 'admin'])->group(function () {
     Route::get('/decks', [DeckController::class, 'index']);
@@ -78,11 +78,9 @@ Route::middleware(['auth:sanctum', 'admin'])->group(function () {
 });
 
 // Admin - Card Routes
-Route::get('/cards', [CardController::class, 'index']);
-Route::post('/cards', [CardController::class, 'store']);
-Route::get('/cards/{card}', [CardController::class, 'show']);
-Route::put('/cards/{card}', [CardController::class, 'update']);
-Route::delete('/cards/{card}', [CardController::class, 'destroy'])->middleware('admin');
+Route::post('/cards', [CardController::class, 'store'])->middleware(['auth:sanctum', 'admin']);
+Route::put('/cards/{card}', [CardController::class, 'update'])->middleware(['auth:sanctum', 'admin']);
+Route::delete('/cards/{card}', [CardController::class, 'destroy'])->middleware(['auth:sanctum', 'admin']);
 
 // Basic Room Routes
 Route::middleware('auth:sanctum')->group(function () {
@@ -97,23 +95,18 @@ Route::middleware('auth:sanctum')->group(function () {
 Route::middleware(['auth:sanctum', 'admin'])->group(function () {
     Route::get('/ship-layouts', [ShipLayoutController::class, 'index']);
     Route::post('/ship-layouts', [ShipLayoutController::class, 'store']);
-    Route::get('/ship-layouts/{shipLayout}', [ShipLayoutController::class, 'show']);
     Route::put('/ship-layouts/{shipLayout}', [ShipLayoutController::class, 'update']);
     Route::delete('/ship-layouts/{shipLayout}', [ShipLayoutController::class, 'destroy']);
 });
 
 // Simulation Log Routes
-Route::post('/simulation-logs', [SimulationLogController::class, 'store'])->middleware('auth:sanctum');
-Route::get('/simulation-logs/{simulationLog}', [SimulationLogController::class, 'show'])->middleware('auth:sanctum');
-Route::get('/simulation-logs/rooms/{roomId}', [SimulationLogController::class, 'getRoomLogs']);
-Route::get('/simulation-logs/rooms/{roomId}/users/{userId}/', [SimulationLogController::class, 'getUserLogs']);
-Route::delete('/simulation-logs/{simulationLog}', [SimulationLogController::class, 'destroy'])->middleware('auth:sanctum', 'admin');
+Route::get('/simulation-logs/rooms/{roomId}/users/{userId}/', [SimulationLogController::class, 'getUserLogs'])->middleware('auth:sanctum', 'user.ownership');
+Route::get('/simulation-logs/rooms/{roomId}', [SimulationLogController::class, 'getRoomLogs'])->middleware('auth:sanctum');
 
 // Market Intelligence Routes
 Route::get('/market-intelligence/deck/{deck}', [MarketIntelligenceController::class, 'forDeck'])->middleware('auth:sanctum');
 Route::post('/market-intelligence/deck/{deck}', [MarketIntelligenceController::class, 'storeOrUpdate'])->middleware('auth:sanctum', 'admin');
 Route::post('/market-intelligence/deck/{deck}/generate-default', [MarketIntelligenceController::class, 'generateDefault'])->middleware('auth:sanctum', 'admin');
-Route::delete('/market-intelligence/{marketIntelligence}', [MarketIntelligenceController::class, 'destroy'])->middleware('auth:sanctum', 'admin');
 
 // Other Room Routes
 Route::middleware(['auth:sanctum', 'room.access'])->group(function () {
@@ -129,35 +122,32 @@ Route::middleware(['auth:sanctum', 'room.access'])->group(function () {
     Route::get('/rooms/{room}/users', [RoomController::class, 'getRoomUsers'])->middleware('auth:sanctum');
     Route::put('/rooms/{room}/set-ports', [RoomController::class, 'setPorts'])->middleware('auth:sanctum', 'admin');
     Route::get('/rooms/{room}/user-port', [RoomController::class, 'getUserPortsV1'])->middleware('auth:sanctum');
-    Route::get('/rooms/{room}/rankings', [RoomController::class, 'getUsersRanking'])->middleware('auth:sanctum');
-    Route::put('/rooms/{room}/swap-config', [RoomController::class, 'updateSwapConfig'])->middleware('auth:sanctum');
+    Route::get('/rooms/{room}/rankings', [RoomController::class, 'getUsersRanking'])->middleware('auth:sanctum', 'admin');
+    Route::put('/rooms/{room}/swap-config', [RoomController::class, 'updateSwapConfig'])->middleware('auth:sanctum', 'admin');
     Route::get('/rooms/{roomId}/port-sequence/{port}', [RoomController::class, 'getProperStackingOrder']);
     Route::get('/rooms/{roomId}/restowage-status', [RoomController::class, 'getRestowageStatus'])->middleware('auth:sanctum');
     Route::get('/rooms/{roomId}/details', [RoomController::class, 'getRoomDetails'])->middleware('auth:sanctum');
 
-    Route::get('/card-temporary/{roomId}/{userId}', [CardTemporaryController::class, 'getCardTemporaries']);
+    Route::get('/card-temporary/{roomId}/{userId}', [CardTemporaryController::class, 'getCardTemporaries'])->middleware('user.ownership');
     Route::get('/card-temporary/all-cards/{roomId}/{deckId}', [CardTemporaryController::class, 'getAllCardTemporaries']);
-    Route::post('/card-temporary/accept', [CardTemporaryController::class, 'acceptCardTemporary']);
-    Route::post('/card-temporary/reject', [CardTemporaryController::class, 'rejectCardTemporary']);
 
     // ShipBay Routes
-    Route::get('/rooms/{room}/ship-bays', [ShipBayController::class, 'index'])->middleware(['auth:sanctum']);
-    Route::post('/rooms/{room}/ship-bays', [ShipBayController::class, 'store'])->middleware(['auth:sanctum']);
-    Route::get('/rooms/{room}/ship-bays/{shipBay}', [ShipBayController::class, 'show'])->middleware(['auth:sanctum']);
+    Route::post('/rooms/{room}/ship-bays', [ShipBayController::class, 'store'])->middleware(['auth:sanctum', 'user.ownership']);
+
 
     // Other ShipBay Routes
-    Route::get('/rooms/{room}/ship-bays/arena-data/{userId}', [ShipBayController::class, 'getConsolidatedArenaData'])->middleware('auth:sanctum');
-    Route::put('/rooms/{room}/ship-bays/{user}/section', [ShipBayController::class, 'updateSection'])->middleware(['auth:sanctum']);
-    Route::post('/rooms/{room}/ship-bays/{user}/moves', [ShipBayController::class, 'incrementMoves'])->middleware('auth:sanctum');
-    Route::post('/rooms/{room}/ship-bays/{user}/cards', [ShipBayController::class, 'incrementCards'])->middleware('auth:sanctum');
+    Route::get('/rooms/{room}/ship-bays/arena-data/{userId}', [ShipBayController::class, 'getConsolidatedArenaData'])->middleware('auth:sanctum', 'user.ownership');
+    Route::put('/rooms/{room}/ship-bays/{user}/section', [ShipBayController::class, 'updateSection'])->middleware(['auth:sanctum', 'user.ownership']);
+    Route::post('/rooms/{room}/ship-bays/{user}/moves', [ShipBayController::class, 'incrementMoves'])->middleware('auth:sanctum, user.ownership');
+    Route::post('/rooms/{room}/ship-bays/{user}/cards', [ShipBayController::class, 'incrementCards'])->middleware('auth:sanctum, user.ownership');
 
     // ShipDock Routes
-    Route::post('/rooms/{room}/ship-docks', [ShipDockController::class, 'store'])->middleware(['auth:sanctum']);
+    Route::post('/rooms/{room}/ship-docks', [ShipDockController::class, 'store'])->middleware(['auth:sanctum', 'user.ownership']);
 
     // Weekly Performance Routes
-    Route::get('/rooms/{roomId}/weekly-performance-all/{userId}', [WeeklyPerformanceController::class, 'getAllWeeklyPerformance']);
-    Route::get('/rooms/{roomId}/ship-bays/financial-summary/{userId}', [WeeklyPerformanceController::class, 'getFinancialSummary']);
+    Route::get('/rooms/{roomId}/weekly-performance-all/{userId}', [WeeklyPerformanceController::class, 'getAllWeeklyPerformance'])->middleware('auth:sanctum', 'user.ownership');
+    Route::get('/rooms/{roomId}/ship-bays/financial-summary/{userId}', [WeeklyPerformanceController::class, 'getFinancialSummary'])->middleware('auth:sanctum', 'user.ownership');
 
     // Capacity Uptake Routes
-    Route::get('/rooms/{roomId}/capacity-uptakes/{userId}/{week?}', [CapacityUptakeController::class, 'getCapacityUptake']);
+    Route::get('/rooms/{roomId}/capacity-uptakes/{userId}/{week?}', [CapacityUptakeController::class, 'getCapacityUptake'])->middleware('auth:sanctum', 'user.ownership');
 });
